pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "llm-gate"
version = "0.2.0"
description = "LLM-Gate — AI-шлюз для обработки входящих инженерных задач"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "jinja2>=3.1.4",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.0",
    "openai>=1.55.0",
    "tiktoken>=0.12.0",
    "faiss-cpu>=1.9.0",
    "sentence-transformers>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[project.scripts]

[tool.hatch.build.targets.wheel]
packages = ["src/app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["src/tests"]
pythonpath = ["src"]
markers = ["slow: marks tests as slow (deselect with '-m \"not slow\"')"]
